abstracted our OpenGL code into seperate game class instead of main.cpp -> fixes clean up order and seg fault
