+++ LearnOpenGL/assets/shaders/fragment.frag
@@ -10,5 +10,5 @@ uniform sampler2D ourTexture;
 void main() {
 //    fragColor = vec4(ourColor, 1.0);
 //    fragColor = ourPos; // negative vals get clamped to 0.0 = black!
-    fragColor = texture(ourTexture, TexCoord);
+    fragColor = texture(ourTexture, TexCoord) * vec4(ourColor, 1.0);
 }
